+++ src/video_core/pica/regs_shader.h
@@ -1,5 +1,6 @@
 // Copyright 2017 Citra Emulator Project
 // Copyright 2024 Borked3DS Emulator Project
+// Copyright 2025 Azahar Emulator Project
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
@@ -24,7 +25,8 @@ struct ShaderRegs {
 
     Common::Vec4<u8> GetIntUniform(u32 index) const {
         const auto& values = int_uniforms[index];
-        return Common::MakeVec<u8>(values.x, values.y, values.z, values.w);
+        return Common::MakeVec<u8>(static_cast<u8>(values.x), static_cast<u8>(values.y),
+                                   static_cast<u8>(values.z), static_cast<u8>(values.w));
     }
 
     INSERT_PADDING_WORDS(0x4);
